+++ ui-components/src/locales/general.json
@@ -909,6 +909,7 @@
   "t.preferNotToSay": "Prefer not to say",
   "t.preferences": "Preferences",
   "t.previous": "Previous",
+  "t.programs": "Programs",
   "t.propertyAmenities": "Property Amenities",
   "t.range": "%{from} to %{to}",
   "t.readLess": "read less",
